requirements.txt
constructs>=10.0.0,<11.0.0
aws-cdk-lib>=2.101.1
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha
python-dotenv>=1.0.0
boto3
ldap3
requests

